Moving LandscapeDemos to repo-builder
