+++ engine/desktop/src/test/java/es/eucm/ead/engine/demos/DemoLauncher.java
@@ -166,11 +166,6 @@ public class DemoLauncher extends JFrame {
 		registerDemo(new BackgroundShadersDemo());
 		registerDemo(new URLImageDemo());
 		registerDemo(new RenderersDemo());
-		registerDemo(new LandscapeDemo());
-		registerDemo(new LandscapeDemo2());
-		registerDemo(new LandscapeDemo3());
-		registerDemo(new LandscapeDemo4());
-		registerDemo(new LandscapeDemo5());
 	}
 
 	private void registerDemo(ExecutableDemoBuilder demoBuilder) {
